+++ src/constants.ts
@@ -1,4 +1,4 @@
-export const VERSION = "0.4.0";
+export const VERSION = "0.4.1";
 
 export const PLUGIN_NAME = "vue-doctor";